consolidate type and typeDefinition computations into function
